requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
cffi==1.15.0
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.3
django-configurations==2.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-widget-tweaks==1.4.12
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycparser==2.21
sqlparse==0.4.2
